Modularize character dashboard columns
